Add quantile_huber_loss
